third_party/WebKit/LayoutTests/inspector/extensions/extensions-reload.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/extensions-test.js"></script>

<script src="resources/test-script.js"></script>
<script type="text/javascript">

window.bar = "foo = " + window.foo;

function extension_testReloadInjectsCode(nextTest)
{
    var valueWithInjectedCode;

    function onPageWithInjectedCodeLoaded()
    {
        webInspector.inspectedWindow.eval("window.bar", function(value) {
            valueWithInjectedCode = value;
            evaluateOnFrontend("InspectorTest.runWhenPageLoads(reply)", onPageWithoutInjectedCodeLoaded);
            webInspector.inspectedWindow.reload();
        });
    }
    function onPageWithoutInjectedCodeLoaded()
    {
        webInspector.inspectedWindow.eval("window.bar", function(value) {
            output("With injected code: " + valueWithInjectedCode);
            output("Without injected code: " + value);
            nextTest();
        });
    }
    evaluateOnFrontend("InspectorTest.runWhenPageLoads(reply)", onPageWithInjectedCodeLoaded);
    webInspector.inspectedWindow.reload({
        injectedScript: "window.foo = 42;"
    });
}

function extension_testReloadPreprocessesCode(nextTest)
{
    evaluateOnFrontend("InspectorTest.runWhenPageLoads(reply)", nextTest);
    function preprocessor(src, url)
    {
        console.log("preprocessor transcoding " + url);
        if(/-test.js/.test(url))
            return src;
        url = url.replace(/\.html/, '_html');
        var preamble = 'console.log("running preprocessed " + \"' + url + '\");\n';
        var sourceURL = '\n//# sourceURL=' + url + '.js\n';
        return preamble + src + sourceURL;
    }
    webInspector.inspectedWindow.reload({
        preprocessingScript: '(' + preprocessor + ')'
    });
}

</script>
</head>
<body onload="runTest()">
<p>Tests that webInspector.inspectedWindow.reload() successfully injects user's code upon reload</p>
</body>
</html>
